mobile/android/tests/browser/chrome/test_selectoraddtab.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1216047
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1216047</title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SpawnTask.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://global/skin"/>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"/>
  <script type="application/javascript" src="head.js"></script>
  <script type="application/javascript">

  "use strict";

  ChromeUtils.import("resource://gre/modules/Services.jsm");
  ChromeUtils.import("resource://gre/modules/Messaging.jsm");
  ChromeUtils.import("resource://gre/modules/Task.jsm");

  // The chrome window
  let chromeWin = Services.wm.getMostRecentWindow("navigator:browser");
  let BrowserApp = chromeWin.BrowserApp;

  // Track the tabs where the tests are happening
  let tabBlank;
  let tabTest;

  function cleanupTabs() {
    if (tabBlank) {
      BrowserApp.closeTab(tabBlank);
      tabBlank = null;
    }

    if (tabTest) {
      BrowserApp.closeTab(tabTest);
      tabTest = null;
    }
  }

  SimpleTest.registerCleanupFunction(function() {
    cleanupTabs();
  });

  const kTestPage = "http://mochi.test:8888/chrome/mobile/android/tests/browser/chrome/basic_article.html";

  add_task(async function test_selectOrAdd() {
    // Add a new tab with a blank page
    tabBlank = BrowserApp.addTab("about:blank", { selected: true, parentId: BrowserApp.selectedTab.id });

    // Now, let's force the target tab to be added
    tabTest = BrowserApp.selectOrAddTab(kTestPage, { selected: true, parentId: BrowserApp.selectedTab.id });
    await promiseBrowserEvent(tabTest.browser, "DOMContentLoaded");

    // Check that basic_article is now selected
    is(BrowserApp.selectedBrowser, tabTest.browser, "Target tab is selected after being added.");

    // Switch back to about:blank
    BrowserApp.selectTab(tabBlank);
    await promiseTabEvent(BrowserApp.deck, "TabSelect");

    // Check that about:blank is selected
    is(BrowserApp.selectedTab, tabBlank, "about:blank is selected.");

    // Use selectOrAddTab to select the existing tab
    BrowserApp.selectOrAddTab(kTestPage, { selected: true, parentId: BrowserApp.selectedTab.id });
    await promiseTabEvent(BrowserApp.deck, "TabSelect");

    // Check that basic_article is now selected
    is(BrowserApp.selectedTab, tabTest, "Target tab is selected.");

    // Switch back to about:blank
    BrowserApp.selectTab(tabBlank);
    await promiseTabEvent(BrowserApp.deck, "TabSelect");

    // Check that about:blank is selected
    is(BrowserApp.selectedTab, tabBlank, "about:blank is selected.");

    // Use selectOrAddTab to select the existing tab using the startsWith flag
    BrowserApp.selectOrAddTab(kTestPage, { selected: true, parentId: BrowserApp.selectedTab.id }, { startsWith: kTestPage });
    await promiseTabEvent(BrowserApp.deck, "TabSelect");

    // Check that basic_article is now selected
    is(BrowserApp.selectedTab, tabTest, "Target tab is selected.");

    // Switch back to about:blank
    BrowserApp.selectTab(tabBlank);
    await promiseTabEvent(BrowserApp.deck, "TabSelect");

    // Check that about:blank is selected
    is(BrowserApp.selectedTab, tabBlank, "about:blank is selected.");

    // Zombify basic_article
    tabTest.zombify();

    // Check that basic_article is actually zombified
    ok(tabTest.browser.__SS_restore, "Target tab is set for delay loading.");
    is(tabTest.browser.currentURI.spec, "about:blank", "Target tab is zombified.");

    // Use selectOrAddTab to select the existing tab
    BrowserApp.selectOrAddTab(kTestPage, { selected: true, parentId: BrowserApp.selectedTab.id });
    await promiseTabEvent(BrowserApp.deck, "TabSelect");

    // Check that basic_article is now selected
    is(BrowserApp.selectedTab, tabTest, "Target tab is selected.");

    cleanupTabs();
  });

  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1216047">Mozilla Bug 1216047</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
